chore: extract common rockcraft options
